requirements.txt
langchain==1.0.2
pdfplumber==0.11.7
pandas==2.3.3
python-docx==1.2.0
pytest==8.4.2
langchain-text-splitters==1.0.0
pinecone==7.3.0
langchain-community==0.4